lfu 与 hashlfu 增加ResizeWeight方法, 可以用来调整lfu引用次数

+++ hashLfu.go
@@ -192,7 +192,7 @@ func (h *HashLfuCache) Resize(size int) (evicted int) {
 
 // ResizeWeight 改变缓存中Weight大小。
 // ResizeWeight 改变缓存中Weight大小。
-func (h *HashLfuCache) ResizeWeight(percentage int64) {
+func (h *HashLfuCache) ResizeWeight(percentage int) {
 	for i := 0; i < h.sliceNum; i++ {
 		h.list[i].lock.Lock()
 		h.list[i].lfu.ResizeWeight(percentage)
